move from shared to config/firebase
