+++ functions/src/logic/getVersionInfoFromUnity.ts
@@ -1,11 +1,12 @@
 import { getDocumentFromUrl } from '../utils/get-document-from-url';
+import { VersionInfo } from '../model/versionInfo';
 
 const UNITY_ARCHIVE_URL = 'https://unity3d.com/get-unity/download/archive';
 
 /**
  * Based on https://github.com/BLaZeKiLL/unity-scraper
  */
-export const getUnityVersionInfo = async () => {
+export const getVersionInfoFromUnity = async (): Promise<VersionInfo[]> => {
   const document = await getDocumentFromUrl(UNITY_ARCHIVE_URL);
 
   const versionInfo = Array.from(
@@ -16,7 +17,7 @@ export const getUnityVersionInfo = async () => {
 
       return {
         version: link.split('/')[0],
-        changeset: link.split('/')[1],
+        changeSet: link.split('/')[1],
       };
     },
   );
